chore: split agent/compact/main.rs into include! files, remove dead code in provider_flow
