+++ src/agent/compact/main.rs
@@ -336,715 +336,8 @@ impl Agent {
     }
 }
 
-pub(crate) fn build_compact_plan(
-    history: &[Message],
-    threshold: usize,
-    force: bool,
-) -> Result<Option<CompactPlan>> {
-    if history.len() < 2 {
-        return Ok(None);
-    }
-
-    let groups = group_history_by_api_round(history);
-    if groups.len() < 2 {
-        return Ok(None);
-    }
-
-    let retained_budget = retained_history_budget(threshold, force);
-    let mut retained_tokens = 0usize;
-    let mut retained_group_index = groups.len() - 1;
-    let latest_group = &groups[retained_group_index];
-    if !force && latest_group.token_estimate > retained_budget {
-        return Ok(Some(CompactPlan {
-            summarize_end: history.len(),
-            retained_start: history.len(),
-        }));
-    }
-
-    for group_index in (1..groups.len()).rev() {
-        let group = &groups[group_index];
-        let next_tokens = retained_tokens.saturating_add(group.token_estimate);
-        if retained_tokens > 0 && next_tokens > retained_budget {
-            break;
-        }
-        retained_tokens = next_tokens;
-        retained_group_index = group_index;
-    }
-
-    let retained_start = groups[retained_group_index].start;
-    if retained_start == 0 {
-        return Ok(Some(CompactPlan {
-            summarize_end: groups[1].start,
-            retained_start: groups[1].start,
-        }));
-    }
-
-    Ok(Some(CompactPlan {
-        summarize_end: retained_start,
-        retained_start,
-    }))
-}
-
-fn retained_history_budget(threshold: usize, force: bool) -> usize {
-    if force {
-        return 1;
-    }
-    threshold
-        .saturating_div(RETAINED_HISTORY_BUDGET_FRACTION)
-        .max(1)
-}
-
-#[allow(dead_code)]
-fn ensure_api_round_boundary_range(history: &[Message], from: usize, up_to: usize) -> Result<()> {
-    let groups = group_history_by_api_round(history);
-    let is_boundary = |idx: usize| {
-        idx == history.len()
-            || groups
-                .iter()
-                .any(|group| group.start == idx || group.end == idx)
-    };
-    if !is_boundary(from) || !is_boundary(up_to) {
-        return Err(anyhow::anyhow!(
-            "partial compaction range must align with API-round boundaries"
-        ));
-    }
-    Ok(())
-}
-
-pub(crate) fn group_history_by_api_round(history: &[Message]) -> Vec<ApiRoundGroup> {
-    let mut groups = Vec::new();
-    let mut group_start = 0usize;
-    let mut current_tokens = 0usize;
-
-    for (idx, message) in history.iter().enumerate() {
-        let starts_new_round = message.role == ROLE_ASSISTANT && idx > group_start;
-        if starts_new_round {
-            debug_assert!(idx > group_start);
-            groups.push(ApiRoundGroup {
-                start: group_start,
-                end: idx,
-                token_estimate: current_tokens.max(1),
-            });
-            group_start = idx;
-            current_tokens = 0;
-        }
-        current_tokens =
-            current_tokens.saturating_add(approximate_token_count_for_message(message));
-    }
-
-    if group_start < history.len() {
-        groups.push(ApiRoundGroup {
-            start: group_start,
-            end: history.len(),
-            token_estimate: current_tokens.max(1),
-        });
-    }
-    debug_assert_eq!(groups.last().map(|group| group.end), Some(history.len()));
-
-    groups
-}
-
-fn build_compact_carry_over(
-    summary: String,
-    compacted_history: &[Message],
-    retrieved_memory_candidates: &[RetrievedMemoryCandidate],
-) -> CompactCarryOver {
-    CompactCarryOver {
-        summary,
-        recent_files: collect_recent_files(compacted_history, RECENT_FILE_CARRY_OVER_LIMIT),
-        recent_file_excerpts: collect_recent_file_excerpts(
-            compacted_history,
-            RECENT_FILE_EXCERPT_LIMIT,
-            RECENT_FILE_EXCERPT_CHAR_LIMIT,
-        ),
-        retrieved_memory: collect_retrieved_memory_carry_over(
-            retrieved_memory_candidates,
-            MEMORY_CARRY_OVER_LIMIT,
-        ),
-        invoked_skills: collect_invoked_skill_carry_over(
-            compacted_history,
-            SKILL_CARRY_OVER_LIMIT,
-            SKILL_INSTRUCTION_PREVIEW_CHAR_LIMIT,
-        ),
-        retained_hooks: collect_retained_context_carry_over(
-            compacted_history,
-            RetainedContextClass::Hooks,
-            HOOK_CARRY_OVER_LIMIT,
-        ),
-        retained_mcp: collect_retained_context_carry_over(
-            compacted_history,
-            RetainedContextClass::Mcp,
-            MCP_CARRY_OVER_LIMIT,
-        ),
-    }
-}
-
-fn build_post_compact_history(
-    before_tokens: usize,
-    carry_over: &CompactCarryOver,
-    retained_history: &[Message],
-) -> Vec<Message> {
-    let mut history = vec![
-        build_compact_boundary_message(before_tokens, carry_over.recent_files.len()),
-        Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!(
-                    "STRUCTURED SUMMARY OF PREVIOUS CONVERSATION:\n{}",
-                    carry_over.summary
-                ),
-                json!({
-                    "type": "compacted_summary",
-                    "text": carry_over.summary,
-                }),
-            ),
-        },
-    ];
-
-    append_post_compact_carry_over(&mut history, carry_over);
-    history.extend_from_slice(retained_history);
-    history
-}
-
-fn append_post_compact_carry_over(history: &mut Vec<Message>, carry_over: &CompactCarryOver) {
-    if !carry_over.recent_files.is_empty() {
-        let recent_files_block = carry_over
-            .recent_files
-            .iter()
-            .map(|path| format!("- {path}"))
-            .collect::<Vec<_>>()
-            .join("\n");
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!(
-                    "RECENT FILES FROM COMPACTED HISTORY:\n{}",
-                    recent_files_block
-                ),
-                json!({
-                    "type": "recent_files",
-                    "files": carry_over.recent_files.clone(),
-                }),
-            ),
-        });
-    }
-
-    if !carry_over.recent_file_excerpts.is_empty() {
-        let excerpt_block = carry_over
-            .recent_file_excerpts
-            .iter()
-            .map(render_recent_file_excerpt)
-            .collect::<Vec<_>>()
-            .join("\n\n");
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!(
-                    "RECENT FILE EXCERPTS FROM COMPACTED HISTORY:\n{}",
-                    excerpt_block
-                ),
-                json!({
-                    "type": "recent_file_excerpts",
-                    "files": carry_over
-                        .recent_file_excerpts
-                        .iter()
-                        .map(recent_file_excerpt_source_item)
-                        .collect::<Vec<_>>(),
-                }),
-            ),
-        });
-    }
-
-    if !carry_over.retrieved_memory.is_empty() {
-        let memory_block = render_retrieved_memory_carry_over(&carry_over.retrieved_memory);
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!("MEMORY CARRY-OVER FROM COMPACTED HISTORY:\n{}", memory_block),
-                json!({
-                    "type": "compaction_carry_over",
-                    "kind": "compacted_memory",
-                    "label": "Memory Carry-over",
-                    "source_descriptor": "history.compaction.memory",
-                    "detail": memory_block,
-                    "inclusion_reason": "carried forward because retrieved memory was available before compaction and should remain available after history replacement",
-                }),
-            ),
-        });
-    }
-
-    if !carry_over.invoked_skills.is_empty() {
-        let skill_block = render_invoked_skill_carry_over(&carry_over.invoked_skills);
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!("SKILL CARRY-OVER FROM COMPACTED HISTORY:\n{}", skill_block),
-                json!({
-                    "type": "compaction_carry_over",
-                    "kind": "compacted_skills",
-                    "label": "Skill Carry-over",
-                    "source_descriptor": "history.compaction.skills",
-                    "detail": skill_block,
-                    "inclusion_reason": "carried forward because these skills were invoked before compaction and may still define the current workflow",
-                }),
-            ),
-        });
-    }
-
-    if !carry_over.retained_hooks.is_empty() {
-        let hook_block = render_retained_context_carry_over(&carry_over.retained_hooks);
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!("HOOK CARRY-OVER FROM COMPACTED HISTORY:\n{}", hook_block),
-                json!({
-                    "type": "compaction_carry_over",
-                    "kind": "compacted_hooks",
-                    "label": "Hook Carry-over",
-                    "source_descriptor": "history.compaction.hooks",
-                    "detail": hook_block,
-                    "inclusion_reason": "carried forward because hook-provided context requested retention before compaction",
-                }),
-            ),
-        });
-    }
-
-    if !carry_over.retained_mcp.is_empty() {
-        let mcp_block = render_retained_context_carry_over(&carry_over.retained_mcp);
-        history.push(Message {
-            role: "system".to_string(),
-            content: compact_source_content(
-                format!("MCP CARRY-OVER FROM COMPACTED HISTORY:\n{}", mcp_block),
-                json!({
-                    "type": "compaction_carry_over",
-                    "kind": "compacted_mcp",
-                    "label": "MCP Carry-over",
-                    "source_descriptor": "history.compaction.mcp",
-                    "detail": mcp_block,
-                    "inclusion_reason": "carried forward because MCP-provided context requested retention before compaction",
-                }),
-            ),
-        });
-    }
-}
-
-fn compact_source_content(text: String, source_item: Value) -> Value {
-    json!([
-        {
-            "type": "text",
-            "text": text,
-        },
-        source_item,
-    ])
-}
-
-pub(crate) fn compaction_summary_timeout() -> Duration {
-    #[cfg(test)]
-    {
-        TEST_COMPACTION_SUMMARY_TIMEOUT
-    }
-    #[cfg(not(test))]
-    {
-        COMPACTION_SUMMARY_TIMEOUT
-    }
-}
-
-fn estimate_history_tokens(history: &[Message]) -> Result<usize> {
-    let bpe = tokenizer()?;
-    history
-        .iter()
-        .map(|message| estimate_message_tokens_with_bpe(message, bpe))
-        .sum::<Result<usize>>()
-}
-
-fn estimate_message_tokens(message: &Message) -> Result<usize> {
-    let bpe = tokenizer()?;
-    estimate_message_tokens_with_bpe(message, bpe)
-}
-
-fn approximate_message_char_count(message: &Message) -> usize {
-    if let Some(text) = message.content.as_str() {
-        text.len()
-    } else {
-        message.content.to_string().len()
-    }
-}
-
-fn approximate_token_count_for_message(message: &Message) -> usize {
-    approximate_message_char_count(message)
-        .saturating_div(4)
-        .max(1)
-}
-
-fn estimate_message_tokens_with_bpe(
-    message: &Message,
-    bpe: &tiktoken_rs::CoreBPE,
-) -> Result<usize> {
-    let rendered;
-    let content = if let Some(text) = message.content.as_str() {
-        text
-    } else {
-        rendered = message.content.to_string();
-        rendered.as_str()
-    };
-    Ok(bpe.encode_with_special_tokens(content).len())
-}
-
-pub(crate) fn tokenizer() -> Result<&'static tiktoken_rs::CoreBPE> {
-    static BPE: OnceLock<std::result::Result<tiktoken_rs::CoreBPE, String>> = OnceLock::new();
-    match BPE.get_or_init(|| tiktoken_rs::cl100k_base().map_err(|err| err.to_string())) {
-        Ok(bpe) => Ok(bpe),
-        Err(err) => Err(anyhow::anyhow!(err.clone())),
-    }
-}
-
-fn collect_recent_files(history: &[Message], limit: usize) -> Vec<String> {
-    let mut collected = Vec::new();
-    for message in history.iter().rev() {
-        if message.role != "assistant" {
-            continue;
-        }
-        let Some(items) = message.content.as_array() else {
-            continue;
-        };
-        for item in items.iter().rev() {
-            if item.get("type").and_then(Value::as_str) != Some("tool_use") {
-                continue;
-            }
-            let Some(tool_name) = item.get("name").and_then(Value::as_str) else {
-                continue;
-            };
-            let Some(input) = item.get("input").and_then(Value::as_object) else {
-                continue;
-            };
-            let Some(path) = input.get("path").and_then(Value::as_str) else {
-                continue;
-            };
-            if !matches!(
-                tool_name,
-                "read_file" | "list_files" | "write_file" | "replace" | "apply_patch"
-            ) {
-                continue;
-            }
-            let normalized = path.replace('\\', "/");
-            if !collected.iter().any(|existing| existing == &normalized) {
-                collected.push(normalized);
-            }
-            if collected.len() >= limit {
-                return collected;
-            }
-        }
-    }
-    collected
-}
-
-fn collect_retrieved_memory_carry_over(
-    candidates: &[RetrievedMemoryCandidate],
-    limit: usize,
-) -> Vec<RetrievedMemoryCandidate> {
-    let mut candidates = candidates
-        .iter()
-        .filter(|candidate| !candidate.detail.trim().is_empty())
-        .cloned()
-        .collect::<Vec<_>>();
-    candidates.sort_by_key(|candidate| candidate.rank);
-    candidates.truncate(limit);
-    candidates
-}
-
-fn render_retrieved_memory_carry_over(candidates: &[RetrievedMemoryCandidate]) -> String {
-    candidates
-        .iter()
-        .map(|candidate| {
-            format!(
-                "- [{}] {}: {}",
-                candidate.kind,
-                candidate.label.trim(),
-                candidate.detail.trim()
-            )
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-fn collect_invoked_skill_carry_over(
-    history: &[Message],
-    limit: usize,
-    instruction_preview_char_limit: usize,
-) -> Vec<InvokedSkillCarryOver> {
-    use std::collections::HashMap;
-
-    let mut pending_invocations = HashMap::<String, SkillInvocationInput>::new();
-    let mut invoked = Vec::<InvokedSkillCarryOver>::new();
-
-    for message in history {
-        match message.role.as_str() {
-            "assistant" => {
-                let Some(items) = message.content.as_array() else {
-                    continue;
-                };
-                for item in items {
-                    if item.get("type").and_then(Value::as_str) != Some("tool_use") {
-                        continue;
-                    }
-                    if item.get("name").and_then(Value::as_str) != Some("skill") {
-                        continue;
-                    }
-                    let Some(tool_use_id) = item.get("id").and_then(Value::as_str) else {
-                        continue;
-                    };
-                    let Some(input) = item.get("input").and_then(Value::as_object) else {
-                        continue;
-                    };
-                    if input.get("action").and_then(Value::as_str) != Some("invoke") {
-                        continue;
-                    }
-                    let Some(skill_name) = input.get("skill_name").and_then(Value::as_str) else {
-                        continue;
-                    };
-                    pending_invocations.insert(
-                        tool_use_id.to_string(),
-                        SkillInvocationInput {
-                            name: skill_name.trim().to_string(),
-                            args: input
-                                .get("args")
-                                .and_then(Value::as_str)
-                                .map(str::trim)
-                                .filter(|value| !value.is_empty())
-                                .map(ToString::to_string),
-                        },
-                    );
-                }
-            }
-            "user" => {
-                let Some(items) = message.content.as_array() else {
-                    continue;
-                };
-                for item in items {
-                    if item.get("type").and_then(Value::as_str) != Some("tool_result") {
-                        continue;
-                    }
-                    if item
-                        .get("is_error")
-                        .and_then(Value::as_bool)
-                        .unwrap_or(false)
-                    {
-                        continue;
-                    }
-                    let Some(tool_use_id) = item.get("tool_use_id").and_then(Value::as_str) else {
-                        continue;
-                    };
-                    let Some(invocation) = pending_invocations.remove(tool_use_id) else {
-                        continue;
-                    };
-                    let result = item
-                        .get("content")
-                        .and_then(Value::as_str)
-                        .and_then(|content| serde_json::from_str::<Value>(content).ok());
-                    invoked.retain(|existing| existing.name != invocation.name);
-                    invoked.push(skill_carry_over_from_result(
-                        invocation,
-                        result.as_ref(),
-                        instruction_preview_char_limit,
-                    ));
-                }
-            }
-            _ => {}
-        }
-    }
-
-    if invoked.len() > limit {
-        invoked = invoked[invoked.len() - limit..].to_vec();
-    }
-    invoked.reverse();
-    invoked
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct SkillInvocationInput {
-    name: String,
-    args: Option<String>,
-}
-
-fn skill_carry_over_from_result(
-    invocation: SkillInvocationInput,
-    result: Option<&Value>,
-    instruction_preview_char_limit: usize,
-) -> InvokedSkillCarryOver {
-    let field = |name: &str| {
-        result
-            .and_then(|value| value.get(name))
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(ToString::to_string)
-    };
-
-    InvokedSkillCarryOver {
-        name: field("name").unwrap_or(invocation.name),
-        title: field("title"),
-        scope: field("scope"),
-        display_path: field("display_path"),
-        args: field("args").or(invocation.args),
-        instruction_preview: field("instructions").map(|instructions| {
-            truncate_excerpt(instructions.as_str(), instruction_preview_char_limit)
-                .trim()
-                .to_string()
-        }),
-    }
-}
-
-fn render_invoked_skill_carry_over(skills: &[InvokedSkillCarryOver]) -> String {
-    skills
-        .iter()
-        .map(|skill| {
-            let mut parts = vec![format!("- {}", skill.name.trim())];
-            if let Some(title) = skill.title.as_deref().filter(|value| !value.is_empty()) {
-                parts.push(format!("  title: {}", title.trim()));
-            }
-            if let Some(scope) = skill.scope.as_deref().filter(|value| !value.is_empty()) {
-                parts.push(format!("  scope: {}", scope.trim()));
-            }
-            if let Some(path) = skill
-                .display_path
-                .as_deref()
-                .filter(|value| !value.is_empty())
-            {
-                parts.push(format!("  path: {}", path.trim()));
-            }
-            if let Some(args) = skill.args.as_deref().filter(|value| !value.is_empty()) {
-                parts.push(format!("  args: {}", args.trim()));
-            }
-            if let Some(preview) = skill
-                .instruction_preview
-                .as_deref()
-                .filter(|value| !value.is_empty())
-            {
-                parts.push(format!("  instructions: {}", preview.trim()));
-            }
-            parts.join("\n")
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RetainedContextClass {
-    Hooks,
-    Mcp,
-}
-
-fn collect_retained_context_carry_over(
-    history: &[Message],
-    class: RetainedContextClass,
-    limit: usize,
-) -> Vec<RetainedContextCarryOver> {
-    let mut retained = Vec::<RetainedContextCarryOver>::new();
-
-    for message in history {
-        for item in message_content_items(&message.content) {
-            let Some(entry) = retained_context_carry_over_from_item(item, class) else {
-                continue;
-            };
-            retained.retain(|existing| existing.source_descriptor != entry.source_descriptor);
-            retained.push(entry);
-        }
-    }
-
-    if retained.len() > limit {
-        retained = retained[retained.len() - limit..].to_vec();
-    }
-    retained.reverse();
-    retained
-}
-
-fn message_content_items(content: &Value) -> Vec<&Value> {
-    if let Some(items) = content.as_array() {
-        return items.iter().collect();
-    }
-    content
-        .as_object()
-        .map(|_| vec![content])
-        .unwrap_or_default()
-}
-
-fn retained_context_carry_over_from_item(
-    item: &Value,
-    class: RetainedContextClass,
-) -> Option<RetainedContextCarryOver> {
-    if item.get("type").and_then(Value::as_str) != Some("compaction_retain_hint") {
-        return None;
-    }
-
-    let source_descriptor = item
-        .get("source_descriptor")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())?;
-    if !retained_context_descriptor_matches(source_descriptor, class) {
-        return None;
-    }
-
-    let detail = item
-        .get("detail")
-        .or_else(|| item.get("text"))
-        .or_else(|| item.get("content"))
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())?;
-    let label = item
-        .get("label")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or(source_descriptor);
-    let inclusion_reason = item
-        .get("inclusion_reason")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToString::to_string);
-
-    Some(RetainedContextCarryOver {
-        label: label.to_string(),
-        source_descriptor: source_descriptor.to_string(),
-        detail: detail.to_string(),
-        inclusion_reason,
-    })
-}
-
-fn retained_context_descriptor_matches(
-    source_descriptor: &str,
-    class: RetainedContextClass,
-) -> bool {
-    match class {
-        RetainedContextClass::Hooks => source_descriptor.starts_with("hook."),
-        RetainedContextClass::Mcp => source_descriptor.starts_with("mcp."),
-    }
-}
-
-fn render_retained_context_carry_over(entries: &[RetainedContextCarryOver]) -> String {
-    entries
-        .iter()
-        .map(|entry| {
-            let mut parts = vec![
-                format!("- {}", entry.label.trim()),
-                format!("  source: {}", entry.source_descriptor.trim()),
-                format!("  detail: {}", entry.detail.trim()),
-            ];
-            if let Some(reason) = entry
-                .inclusion_reason
-                .as_deref()
-                .filter(|value| !value.is_empty())
-            {
-                parts.push(format!("  reason: {}", reason.trim()));
-            }
-            parts.join("\n")
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
+include!("compact_a.rs");
+include!("compact_b.rs");
 fn collect_recent_file_excerpts(
     history: &[Message],
     limit: usize,
